+++ src/quacc/recipes/gulp/core.py
@@ -11,6 +11,7 @@ if TYPE_CHECKING:
     from ase.atoms import Atoms
 
     from quacc.schemas._aliases.ase import RunSchema
+    from quacc.utils.files import Filenames, SourceDirectory
 
 
 @job
@@ -20,6 +21,7 @@ def static_job(
     keywords: list[str] | None = None,
     options: list[str] | None = None,
     library: str | None = None,
+    copy_files: SourceDirectory | dict[SourceDirectory, Filenames] | None = None,
 ) -> RunSchema:
     """
     Carry out a single-point calculation.
@@ -40,6 +42,8 @@ def static_job(
         available keys, refer to the `ase.calculators.gulp.GULP` calculator.
     library
         Filename of the potential library file, if required.
+    copy_files
+        Files to copy (and decompress) from source to the runtime directory.
 
     Returns
     -------
@@ -59,6 +63,7 @@ def static_job(
         keyword_swaps=keywords,
         option_swaps=options,
         additional_fields={"name": "GULP Static"},
+        copy_files=copy_files,
     )
 
 
@@ -70,6 +75,7 @@ def relax_job(
     keywords: list[str] | None = None,
     options: list[str] | None = None,
     library: str | None = None,
+    copy_files: SourceDirectory | dict[SourceDirectory, Filenames] | None = None,
 ) -> RunSchema:
     """
     Carry out a structure relaxation.
@@ -92,6 +98,8 @@ def relax_job(
         available keys, refer to the `ase.calculators.gulp.GULP` calculator.
     library
         Filename of the potential library file, if required.
+    copy_files
+        Files to copy (and decompress) from source to the runtime directory.
 
     Returns
     -------
@@ -114,4 +122,5 @@ def relax_job(
         keyword_swaps=keywords,
         option_swaps=options,
         additional_fields={"name": "GULP Relax"},
+        copy_files=copy_files,
     )