+++ src/quacc/runners/ase.py
@@ -38,6 +38,8 @@ if TYPE_CHECKING:
     from ase.md.md import MolecularDynamics
     from ase.optimize.optimize import Optimizer
 
+    from quacc.utils.files import Filenames, SourceDirectory
+
     class OptimizerKwargs(TypedDict, total=False):
         """
         Type hint for `optimizer_kwargs` in [quacc.runners.ase.run_opt][].
@@ -59,7 +61,7 @@ if TYPE_CHECKING:
 def run_calc(
     atoms: Atoms,
     geom_file: str | None = None,
-    copy_files: str | Path | list[str | Path] | None = None,
+    copy_files: SourceDirectory | dict[SourceDirectory, Filenames] | None = None,
     get_forces: bool = False,
 ) -> Atoms:
     """
@@ -81,7 +83,7 @@ def run_calc(
         atoms.get_potential_energy() function to update the positions, as this
         varies between codes.
     copy_files
-        Filenames to copy from source to scratch directory.
+        Files to copy (and decompress) from source to the runtime directory.
     get_forces
         Whether to use `atoms.get_forces()` instead of `atoms.get_potential_energy()`.
 
@@ -141,7 +143,7 @@ def run_opt(
     optimizer_kwargs: OptimizerKwargs | None = None,
     store_intermediate_files: bool = False,
     run_kwargs: dict[str, Any] | None = None,
-    copy_files: str | Path | list[str | Path] | None = None,
+    copy_files: SourceDirectory | dict[SourceDirectory, Filenames] | None = None,
 ) -> Optimizer:
     """
     Run an ASE-based optimization in a scratch directory and copy the results back to
@@ -174,7 +176,7 @@ def run_opt(
     run_kwargs
         Dictionary of kwargs for the run() method of the optimizer.
     copy_files
-        Filenames to copy from source to scratch directory.
+        Files to copy (and decompress) from source to the runtime directory.
 
     Returns
     -------
@@ -375,7 +377,7 @@ def run_md(
 def run_vib(
     atoms: Atoms,
     vib_kwargs: VibKwargs | None = None,
-    copy_files: str | Path | list[str | Path] | None = None,
+    copy_files: SourceDirectory | dict[SourceDirectory, Filenames] | None = None,
 ) -> Vibrations:
     """
     Run an ASE-based vibration analysis in a scratch directory and copy the results back
@@ -392,7 +394,7 @@ def run_vib(
     vib_kwargs
         Dictionary of kwargs for the [ase.vibrations.Vibrations][] class.
     copy_files
-        Filenames to copy from source to scratch directory.
+        Files to copy (and decompress) from source to the runtime directory.
 
     Returns
     -------