+++ tests/core/calculators/vasp/test_vasp.py
@@ -1,6 +1,7 @@
 import os
 from copy import deepcopy
 from pathlib import Path
+from shutil import which
 
 import numpy as np
 import pytest
@@ -9,12 +10,14 @@ from ase.calculators.singlepoint import SinglePointDFTCalculator
 from ase.calculators.vasp import Vasp as Vasp_
 from ase.constraints import FixAtoms, FixBondLength
 from ase.io import read
+from pymatgen.io.vasp.sets import MPRelaxSet, MPScanRelaxSet
 
 from quacc import SETTINGS
 from quacc.calculators.vasp import Vasp, presets
 from quacc.schemas.prep import prep_magmoms, prep_next_run
 
 FILE_DIR = Path(__file__).parent
+PSEUDO_DIR = FILE_DIR / "fake_pseudos"
 
 DEFAULT_SETTINGS = SETTINGS.model_copy()
 
@@ -69,9 +72,9 @@ def test_presets():
     assert calc.exp_params["ediff"] == 1e-5
     assert calc.float_params["encut"] == 450
 
-    calc = Vasp(atoms, xc="scan", preset="MPScanSet")
+    calc = Vasp(atoms, xc="scan", pmg_input_set=MPScanRelaxSet)
     assert calc.xc.lower() == "scan"
-    assert calc.string_params["algo"] == "all"
+    assert calc.string_params["algo"].lower() == "all"
     assert calc.exp_params["ediff"] == 1e-5
 
 
@@ -164,6 +167,7 @@ def test_magmoms(atoms_mag, atoms_nomag, atoms_nospin):
     atoms[-1].symbol = "Fe"
     calc = Vasp(atoms, preset="BulkSet")
     atoms.calc = calc
+    assert atoms.get_chemical_symbols() == ["Cu", "Cu", "Cu", "Fe"]
     assert atoms.get_initial_magnetic_moments().tolist() == [2.0] * (len(atoms) - 1) + [
         5.0
     ]
@@ -186,16 +190,15 @@ def test_magmoms(atoms_mag, atoms_nomag, atoms_nospin):
 
     atoms = bulk("Cu") * (2, 2, 1)
     atoms[-1].symbol = "Fe"
-    calc = Vasp(atoms, preset="MPScanSet")
+    calc = Vasp(atoms, pmg_input_set=MPScanRelaxSet)
     atoms.calc = calc
-    assert atoms.get_initial_magnetic_moments().tolist() == [1.0] * (len(atoms) - 1) + [
-        5.0
-    ]
+    assert atoms.get_chemical_symbols() == ["Cu", "Cu", "Cu", "Fe"]
+    assert calc.parameters["magmom"] == [0.6, 0.6, 0.6, 5.0]
 
     atoms = bulk("Cu") * (2, 2, 1)
     atoms[-1].symbol = "Fe"
     atoms.set_initial_magnetic_moments([3.14] * (len(atoms) - 1) + [1.0])
-    calc = Vasp(atoms, preset="BulkSet")
+    calc = Vasp(atoms, pmg_input_set=MPScanRelaxSet)
     atoms.calc = calc
     assert atoms.get_initial_magnetic_moments().tolist() == [3.14] * (
         len(atoms) - 1
@@ -707,7 +710,7 @@ def test_setups():
     assert calc.parameters["setups"]["Cu"] == ""
 
     atoms = bulk("Cu")
-    calc = Vasp(atoms, preset="MPScanSet")
+    calc = Vasp(atoms, pmg_input_set=MPScanRelaxSet)
     assert calc.parameters["setups"]["Cu"] == "_pv"
 
     atoms = bulk("Cu")
@@ -719,7 +722,12 @@ def test_setups():
     assert calc.parameters["setups"]["Cu"] == ""
 
     atoms = bulk("Cu")
-    calc = Vasp(atoms, setups="minimal", preset="MPScanSet")
+    calc = Vasp(atoms, setups="minimal", preset="BulkSet")
+    assert isinstance(calc.parameters["setups"], str)
+    assert calc.parameters["setups"] == "minimal"
+
+    atoms = bulk("Cu")
+    calc = Vasp(atoms, setups="minimal", pmg_input_set=MPScanRelaxSet)
     assert isinstance(calc.parameters["setups"], str)
     assert calc.parameters["setups"] == "minimal"
 
@@ -823,3 +831,80 @@ def test_preset_override():
 
     calc = Vasp(atoms, preset="BulkSet", efermi=None)
     assert calc.parameters.get("efermi") is None
+
+
+def test_pmg_input_set():
+    atoms = bulk("Cu")
+    calc = Vasp(atoms, pmg_input_set=MPRelaxSet, incar_copilot="off")
+    assert calc.parameters == {
+        "algo": "fast",
+        "ediff": 5e-05,
+        "encut": 520,
+        "ibrion": 2,
+        "isif": 3,
+        "ismear": -5,
+        "ispin": 2,
+        "lasph": True,
+        "lorbit": 11,
+        "lreal": "auto",
+        "lwave": False,
+        "nelm": 100,
+        "nsw": 99,
+        "pp": "pbe",
+        "prec": "accurate",
+        "sigma": 0.05,
+        "magmom": [0.6],
+        "lmaxmix": 4,
+        "kpts": [11, 11, 11],
+        "gamma": True,
+        "setups": {"Cu": "_pv"},
+    }
+
+
+def test_pmg_input_set2():
+    atoms = bulk("Fe") * (2, 1, 1)
+    atoms[0].symbol = "O"
+    calc = Vasp(atoms, pmg_input_set=MPRelaxSet, incar_copilot="off")
+    assert calc.parameters == {
+        "algo": "fast",
+        "ediff": 0.0001,
+        "encut": 520,
+        "ibrion": 2,
+        "isif": 3,
+        "ismear": -5,
+        "ispin": 2,
+        "lasph": True,
+        "ldau": True,
+        "ldauj": [0, 0],
+        "ldaul": [0, 2.0],
+        "ldautype": 2,
+        "ldauu": [0, 5.3],
+        "ldauprint": 1,
+        "lorbit": 11,
+        "lreal": "auto",
+        "lwave": False,
+        "nelm": 100,
+        "nsw": 99,
+        "pp": "pbe",
+        "prec": "accurate",
+        "sigma": 0.05,
+        "magmom": [2.3, 2.3],
+        "lmaxmix": 4,
+        "kpts": [5, 11, 11],
+        "gamma": True,
+        "setups": {"Fe": "_pv", "O": ""},
+    }
+
+
+@pytest.mark.skipif(which(SETTINGS.VASP_CMD), reason="VASP is installed")
+def test_run(monkeypatch, tmp_path):
+    monkeypatch.chdir(tmp_path)
+
+    atoms = bulk("Cu")
+    calc = Vasp(atoms, xc="PBE", use_custodian=False)
+    assert calc._run() > 0
+
+    atoms = bulk("Cu")
+    calc = Vasp(atoms, xc="PBE", use_custodian=True)
+    with pytest.raises(FileNotFoundError):
+        calc._run()